[server] | refactor(video dto): 모듈 export하는 index.ts 삭제
