--- packages/server/src/video/dto/index.ts
@@ -1,3 +1,0 @@
-export { VideoResponseDto } from './video-response.dto';
-export { VideoUserDto } from './video-user.dto';
-export { VideosQueryDto } from './videos-query.dto';
